pyproject.toml
[project]
name = "flow-metrics"
version = "0.1.0"
description = "API client for collecting and analyzing hip-hop artist vocabulary data"
authors = [
    { name = "Brett Plemons", email = "54540840+PlemonsBrett@users.noreply.github.com" },
]
readme = "README.md"
requires-python = ">=3.12"

[tool.poetry]
packages = [{ include = "flow_metrics" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
requests = "^2.32.3"
pydantic = "^2.11.3"
python-dotenv = "^1.1.0"
spotipy = "^2.25.1"
nltk = "^3.9.1"
pandas = "^2.2.3"
rich = "^14.0.0"
pydantic-settings = "^2.9.1"
beautifulsoup4 = "^4.13.4"
lyricsgenius = "^3.6.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"
ruff = "^0.11.7"
pre-commit = "^4.2.0"
isort = "^6.0.1"
pyright = "^1.1.400"
pytest-mock = "^3.14.0"
vcrpy = "^7.0.0"
responses = "^0.25.7"

[tool.ruff]
# Enable pycodestyle (E), Pyflakes (F), isort (I)
select = [
    "E",
    "F",
    "I",
    "N",
    "UP",
    "ANN",
    "B",
    "A",
    "COM",
    "C4",
    "DTZ",
    "T10",
    "T20",
    "PT",
    "Q",
    "RET",
    "SIM",
    "ARG",
    "ERA",
]
ignore = ["ANN101", "ANN102"] # Ignore self and cls annotation requirements
line-length = 100
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]           # Ignore unused imports in __init__.py files
"tests/*" = ["ANN", "ARG", "S101"] # Ignore annotations in tests

[tool.pyright]
include = ["flow_metrics"]
exclude = ["**/node_modules", "**/__pycache__", "**/.venv"]
typeCheckingMode = "strict"
reportMissingIMports = true
reportMissingTypeStubs = false
pythonVersion = "3.12"
pythonPlaftorm = "All"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=flow_metrics --cov-report=term-missing --cov-report=xml"
